Swapping in refactored store (#326)
