+++ cypress/unit/store-action-processDataTable.cy.js
@@ -1,4 +1,4 @@
-import { actions, getters, mutations } from "~/store/index-refactor";
+import { actions, getters, mutations } from "~/store";
 
 const context = {
     commit: () => {},
